feat: add validation

+++ app/src/main/java/com/epam/esm/model/service/impl/GiftCertificateServiceImpl.java
@@ -6,6 +6,7 @@ import com.epam.esm.model.dao.GiftCertificateDao;
 import com.epam.esm.model.dao.QueryCustomizer;
 import com.epam.esm.model.service.GiftCertificateService;
 import com.epam.esm.model.service.ServiceException;
+import com.epam.esm.model.service.validator.EntityValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,6 +22,10 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
 
     @Override
     public long add(GiftCertificate giftCertificate) throws ServiceException {
+        if (giftCertificate == null) {
+            throw new IllegalArgumentException("The supplied [GiftCertificate] is " +
+                    "required and must not be null");
+        }
         try {
             return giftCertificateDao.add(giftCertificate);
         } catch (DaoException e) {
@@ -30,6 +35,9 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
 
     @Override
     public Optional<GiftCertificate> findById(long id) throws ServiceException {
+        if (!EntityValidator.isIdValid(id)) {
+            throw new IllegalArgumentException("Id must be positive");
+        }
         try {
             return giftCertificateDao.findById(id);
         } catch (DaoException e) {
@@ -44,6 +52,10 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
 
     @Override
     public List<GiftCertificate> findAll(QueryCustomizer queryCustomizer) throws ServiceException {
+        if (queryCustomizer == null) {
+            throw new IllegalArgumentException("The supplied [QueryCustomizer] is " +
+                    "required and must not be null");
+        }
         try {
             return giftCertificateDao.findAll(queryCustomizer);
         } catch (DaoException e) {
@@ -53,6 +65,13 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
 
     @Override
     public void update(long id, GiftCertificate giftCertificate) throws ServiceException {
+        if (!EntityValidator.isIdValid(id)) {
+            throw new IllegalArgumentException("Id must be positive");
+        }
+        if (giftCertificate == null) {
+            throw new IllegalArgumentException("The supplied [GiftCertificate] is " +
+                    "required and must not be null");
+        }
         try {
             giftCertificateDao.update(id, giftCertificate);
         } catch (DaoException e) {
@@ -62,6 +81,9 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
 
     @Override
     public void delete(long id) throws ServiceException {
+        if (!EntityValidator.isIdValid(id)) {
+            throw new IllegalArgumentException("Id must be positive");
+        }
         try {
             giftCertificateDao.delete(id);
         } catch (DaoException e) {
@@ -71,11 +93,23 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
 
     @Override
     public List<GiftCertificate> findByTagName(String tagName) throws ServiceException {
+        if (tagName == null) {
+            throw new IllegalArgumentException("The supplied [String] is " +
+                    "required and must not be null");
+        }
         return findByTagName(tagName, new QueryCustomizer());
     }
 
     @Override
     public List<GiftCertificate> findByTagName(String tagName, QueryCustomizer queryCustomizer) throws ServiceException {
+        if (tagName == null) {
+            throw new IllegalArgumentException("The supplied [String] is " +
+                    "required and must not be null");
+        }
+        if (queryCustomizer == null) {
+            throw new IllegalArgumentException("The supplied [QueryCustomizer] is " +
+                    "required and must not be null");
+        }
         try {
             return giftCertificateDao.findByTagName(tagName, queryCustomizer);
         } catch (DaoException e) {
